+++ src/main/java/com/zuzihe/slackbot/slack/http/event/SlackEventRouter.java
@@ -1,8 +1,8 @@
-package com.zuzihe.slackbot.slack.http;
+package com.zuzihe.slackbot.slack.http.event;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.zuzihe.slackbot.message.infra.GeminiService;
-import com.zuzihe.slackbot.slack.http.infra.SlackWebClient;
+import com.zuzihe.slackbot.slack.http.global.infra.SlackWebClient;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
