+++ src/main/java/com/zuzihe/slackbot/slack/http/install/app/service/StateService.java
@@ -1,4 +1,4 @@
-package com.zuzihe.slackbot.install.app.service;
+package com.zuzihe.slackbot.slack.http.install.app.service;
 
 import org.springframework.stereotype.Service;
 import java.security.SecureRandom;
