+++ src/main/java/com/zuzihe/slackbot/slack/http/link/account/service/JwtService.java
@@ -1,4 +1,4 @@
-package com.zuzihe.slackbot.link.account.service;
+package com.zuzihe.slackbot.slack.http.link.account.service;
 
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.Keys;
