+++ src/main/java/com/zuzihe/slackbot/slack/http/link/account/service/SlackUserMappingService.java
@@ -1,4 +1,4 @@
-package com.zuzihe.slackbot.link.account.service;
+package com.zuzihe.slackbot.slack.http.link.account.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
